Text input, material changing, setup for duplicating entities
Added a text input GUI as well as a container GUI that can hold multiple GUIs.
Can now load new materials from object files and apply them via text input, cannot use other types of files.
Setup for duplicating entities. DrawMeshData also exists for entities so they can be edited.
A lot of suboptimal changes have been made that may be changed later.
I believe that I edited the logic for the selected entity color, but it's been too long and I don't remember what exactly is different, but it's probably slightly improved.
Also I think there used to be an issue with the skybox's texture messing with the textures and materials in the cache resulting in objects having the wrong texture/material, but that seems to be resolved somehow. Again too long for me to remember how.