Entity Duplication
Entities can now be duplicated. Not fully tested. To get this to work properly had to save list of models rather than converting map to list to preserve order of lists. Also ability to move objects was added earlier, but I forgot to put it into the update I did that in.