Added Entity List, made GuiContainer more dynamic, added Entity visibility
All entities are displayed on entity list and can be selected, made invisible, and duplicated from it.
GuiContainer now uses a list to determine what GUIs to draw. To accommodate this and TextCheck's need to have access to the renderer, IGuiInstance has been updated to allow this change. LightControls and DropFileLoadType updated as well.
In order to get the symbol used for visibility on the entity list, had to add load font symbol, while doing this, decided to change font of GUI as a whole.
Entities can now be made invisible. Selecting entities ignores invisible entities. Hidden entities are still selected if made hidden and can be selectable from entity list.
Jar files have been updated