DropFileLoadType is now a thread with some bugs along with it
Bugs:
File name is not decoded correctly sometimes.
Unsure why crashing sometimes (probably threading issues, but not sure).
Animated meshes not animating correctly.

Other Changes:
Moved callbacks so they're not constantly being put in place.
DropFileLoadType now just opens an IMGUI window rather than itself in a new GLFW window (originally I just wanted to see if it was possible to do this, the way it is now is how I envisioned it actually being implemented).
I don't remember if meshmode was mentioned in a previous thing, but it's not been implemented yet. The idea is to interact with the individual meshdatas that make up a model/entity.