Should now be able to load in Animated meshes
Title says part of it. There is what I believe to be either an error in the original code that I copied or I have made changes that changed the logic that was used then. Either way, it had to do with how animation data was stored in buffers.
Added semaphore to protect parallelization to avoid concurrent editing. There exists an error where the file dropped will not be decoded correctly. Unsure what causes this.